update param to u32 insted of u8

+++ btrblocks/btrblocks.cpp
@@ -11,7 +11,7 @@ void BtrBlocksConfig::configure(const std::function<void(BtrBlocksConfig&)>& f) 
 }
 
 
-void configure_btrblocks(uint8_t max_depth) {
+void configure_btrblocks(uint32_t max_depth) {
     BtrBlocksConfig::configure([&](BtrBlocksConfig &config){
         std::cout << "setting max cascade depth to (" << max_depth  << ")" << std::endl;
         config.integers.max_cascade_depth = max_depth;
