+++ btrblocks/btrblocks.hpp
@@ -22,7 +22,7 @@ enum class SchemeSelection : uint8_t { SAMPLE, TRY_ALL };
 
 void emretestfunc(int32_t value);
 
-void configure_btrblocks(uint8_t max_depth);
+void configure_btrblocks(uint32_t max_depth);
 
 struct BtrBlocksConfig {
   // clang-format off
